Upgrade to using @fluentui/react
Updated the package from using office-ui-fabric-react to fluentui/react
so that newer controls can be used and we are not developing with a
deprecated library from the outset. There was only one breaking change
that was discovered, a reference to SCSS files from the legacy package
in ApiCalculator.module.scss.