AIP Calculator Webpart Settings (#2)
* Add new localization strings for properties.

Added new localization strings for the the webpart properties flyout
when it is being configured. The strings that were added were:
- RolesFieldLabel: The heading for the roles field
- RolesFieldDescription: The description of the roles field
- IndividualPerformanceBandsFieldLabel: The heading of the individual
  performance bands field
- IndividualPerformanceBandsFieldDescription: The description of the
  individual performance bands field
- CompanyPerformanceBandsFieldLabel: The heading of the company
  performance bands field
- CompanyPerformanceBandsFieldDescription: The description of the
  company performance bands field

* Add classes to represent dropdown options.

Added `AipDropdown` as a base class to represent a dropdown option with
the preset default values.

Added `RoleDropdownOption` to represent a dropdown option for roles.
Added `CompanyPerformanceBandDropdownOption` to represent a dropdown
option for company performance bands. Added
`IndividualPerformanceBandDropdownOption` to represent a dropdown option
for individual performance bands. All these classes inherit from
`AipDropdown`.

* Update webpart properties to show new fields.

Added the ability to modify the webpart the newly created properties:
- `roles`
- `individualPerformanceBands`
- `companyPerformanceBands`
These are multiline text fields that will contain CSV like data to be
then transformed into dropdown options for the end user.

* Add createDropDownOptions method.

The createDropDownOptions method will allow the CSV like data entered in
the webpart properties to be transformed into `AipDropdownOption`
objects.

* Add members to IAipCalculatorProps.ts

This commit adds members to `IAipCalculatorProps` and also injects the
`AipDrowdownOption` objects via the `render` method in
`AipCalculatorWebPart`. The `AipCalculatorWebPart` render method was
included to maintain the integrity of the build.